Extract tests for Connector and System Access Control to base test
